Move and rename UIUtils
