+++ packages/pmml-editor/src/editor/components/mocked/MockTitle.tsx
@@ -13,5 +13,12 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
+import * as React from "react";
 
-export * from "./PMMLUtils";
+interface TitleProps {
+  title: string;
+}
+
+export const Title = (props: TitleProps) => {
+  return <h1 className="pf-c-title pf-m-xl">{props.title}</h1>;
+};
