+++ packages/pmml-editor/src/editor/components/mocked/MockVersionUI.tsx
@@ -17,7 +17,7 @@ import * as React from "react";
 import { CSSProperties } from "react";
 import { useDispatch, useSelector } from "react-redux";
 import { Actions } from "../../reducers";
-import { Timestamp, Title } from "../../utils";
+import { Timestamp, Title } from ".";
 import { PMML } from "@kogito-tooling/pmml-editor-marshaller";
 
 const style: CSSProperties = {
